Refactor to block-by-block (de)compressing
yeap what a nice day to be alive! read the file block-by-block and
calculate them and write to the output image directly. No full image
calculating, which was very inefficent.

even, it's still slow!
callgrind says that 'dct_block()' takes 87% of the running time..
So, optimization to DCT calcuation is required!!
But how to do??